Widen scaler-cached IVS row math and fix the Skia benchmark width buffer
TryGetDeltaWithScalers (the per-clone scaler-cached HVAR/VVAR path) had
the same `innerIndex * rowBytes` int overflow that TryGetDelta did, and
is reachable from the same public metrics APIs on a varied clone. Widen
its row offset to long before the bounds check.

The Skia baseline benchmarks reinterpreted a 200-element ushort advance
span as floats via MemoryMarshal.Cast, which halves the element count -
so GetGlyphWidths received a 100-float buffer for 200 glyphs (saved only
by the oversized backing array). Give Skia a real one-float-per-glyph
buffer so the "within 2x of Skia" comparison measures a well-formed call.

+++ src/Avalonia.Base/Media/Fonts/Tables/Variation/ItemVariationStore.cs
@@ -219,7 +219,11 @@ namespace Avalonia.Media.Fonts.Tables.Variation
 
             var regionIndexesStart = subtableStart + 6;
             var deltaDataStart = regionIndexesStart + regionIndexCount * 2;
-            var rowStart = deltaDataStart + innerIndex * rowBytes;
+
+            // Widen to long before the bounds check: innerIndex × rowBytes overflows int (see
+            // TryGetDelta) — the scaler-cached path is reachable from the same public metrics
+            // APIs on a varied clone, so it needs the identical guard.
+            var rowStart = deltaDataStart + innerIndex * (long)rowBytes;
 
             if (rowStart + rowBytes > span.Length)
             {
@@ -229,7 +233,7 @@ namespace Avalonia.Media.Fonts.Tables.Variation
             // Same row walk as TryGetDelta, but the per-region scaler comes from the
             // precomputed array instead of being computed from activeCoords.
             var sum = 0f;
-            var bytePos = rowStart;
+            var bytePos = (int)rowStart;
 
             for (var r = 0; r < regionIndexCount; r++)
             {
